+++ src/events/chghost.h
@@ -1,0 +1,8 @@
+#ifndef SRC_EVENTS_CHGHOST_H_
+#define SRC_EVENTS_CHGHOST_H_
+
+__SWIRC_BEGIN_DECLS
+void event_chghost(struct irc_message_compo *);
+__SWIRC_END_DECLS
+
+#endif